Rename src/traceLog.ts to src/protocol.ts
Part of ruby/vscode-rdbg#174

This PR changed the file name for easier understanding.
